requirements.txt
fastapi==0.104.1
python-dotenv==1.0.0
openai==1.3.3
twilio==9.3.6
supabase==2.0.3
pinecone-client==3.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-multipart==0.0.6
uvicorn==0.24.0